refactor: gnb
