--- renderer/features/LNB/stores/lnbOpenAtom.ts
@@ -1,4 +1,0 @@
-import { atomWithLocalStorage } from '@/stores/utils/atomWithLocalStorage';
-
-export const lnbOpenAtom = atomWithLocalStorage('lnbOpen', true);
-lnbOpenAtom.debugLabel = 'lnbOpen';
