+++ renderer/features/settings/components/shortcut/ShortCutInputOverlay.tsx
