+++ renderer/pages/tools/TimerPage.tsx
@@ -3,7 +3,7 @@ import ActiveTimer from '@/features/timer/components/ActiveTimer';
 
 const TimerPage = () => {
   return (
-    <div className={'flex h-full px-4 pb-4'}>
+    <div className={'flex h-full px-1 pb-1'}>
       <section className={'flex flex-1 items-center justify-center'}>
         <ActiveTimer />
       </section>
